packages/sentinel-core/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sentinel-core"
version = "0.1.0"
description = "Platform wrapper, auth, bus bridge, and reverse proxy for SENTINEL"
requires-python = ">=3.11"
dependencies = [
    "sentinel-common",
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "httpx>=0.27",
    "websockets>=12.0",
    "pyjwt>=2.8",
    "bcrypt>=4.1",
    "pydantic-settings>=2.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-httpx>=0.30",
    "fakeredis>=2.21",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.hatch.build.targets.wheel]
packages = ["sentinel_core"]
